Pipe Puzzle Minigame
Based off the Bioshock version.
Call with pbPipePuzzle(level)
Define new levels in the LevelLayouts array.
Each level is a 5x5 square, though in theory the script should perform with other sizes if the functionality were readded.
Pipes can fill in both directions.
pbPipePuzzle returns true on sucess and false on quit and fail (from overflow)
The start and end pipes need to be in the background image.